Buildable Tiles
Creates some scripts for building on tiles.

Currently needs BuildableTile attached a component to a tile that can be
built on.

A cube is currently the place holder for interacting with, but this will
be changed a transparent cube that matches the size of the tile.

BuildMenu doesn't bring up the menu yet, but the code will occur in
there that creates towers.

Added a script the player where they can hit E to interact with objects.